typo corrected
We started our strategy from the bottom row, and at each step, we considered the following: if we are positioned at an element in the second-to-last row, we have two possible numbers to choose to complete our path. Therefore, we will definitely choose the larger one to get a greater sum.